운동기록 요청 수정

+++ src/main/java/burnCalories/diet/DTO/userDTO/exerciseLog/ResponseDurationLogDTO.java
@@ -9,7 +9,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class ResponseDurationLogDTO {
     private String exerciseType;
-    private float duration;
+    private double duration;
     private LocalDateTime startTime;
     private LocalDateTime endTime;
 }
